+++ components/Footers/FooterAdmin.js
@@ -23,7 +23,7 @@ export default function FooterAdmin() {
                 <li>
                   <a
                     href="https://www.creative-tim.com?ref=nnjs-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    className="text-blueGray-600 hover:text-brand-purple text-sm font-semibold block py-1 px-3"
                   >
                     Creative Tim
                   </a>
@@ -31,7 +31,7 @@ export default function FooterAdmin() {
                 <li>
                   <a
                     href="https://www.creative-tim.com/presentation?ref=nnjs-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    className="text-blueGray-600 hover:text-brand-purple text-sm font-semibold block py-1 px-3"
                   >
                     About Us
                   </a>
@@ -39,7 +39,7 @@ export default function FooterAdmin() {
                 <li>
                   <a
                     href="http://blog.creative-tim.com?ref=nnjs-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    className="text-blueGray-600 hover:text-brand-purple text-sm font-semibold block py-1 px-3"
                   >
                     Blog
                   </a>
@@ -47,7 +47,7 @@ export default function FooterAdmin() {
                 <li>
                   <a
                     href="https://github.com/creativetimofficial/notus-nextjs/blob/main/LICENSE.md?ref=nnjs-footer-admin"
-                    className="text-blueGray-600 hover:text-blueGray-800 text-sm font-semibold block py-1 px-3"
+                    className="text-blueGray-600 hover:text-brand-purple text-sm font-semibold block py-1 px-3"
                   >
                     MIT License
                   </a>