+++ pages/auth/forgetpassword.js
@@ -6,7 +6,7 @@ import { useRouter } from "next/router";
 
 import Auth from "layouts/Auth.js";
 
-export default function Login() {
+export default function ForgetPassword() {
   const router = useRouter();
 
   const handleUserLogin = async (e) => {
@@ -27,41 +27,17 @@ export default function Login() {
                       className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
                       htmlFor="grid-password"
                     >
-                      Phone number
+                     Email
                     </label>
-
                     <input
                       type="number"
                       className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                      placeholder="Phone number"
+                      placeholder="email"
                     />
                   </div>
 
-                  <div className="relative w-full mb-3">
-                    <label
-                      className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
-                      htmlFor="grid-password"
-                    >
-                      Password
-                    </label>
-                    <input
-                      type="password"
-                      className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                      placeholder="Password"
-                    />
-                  </div>
-                  <div>
-                    <label className="inline-flex items-center cursor-pointer">
-                      <input
-                        id="customCheckLogin"
-                        type="checkbox"
-                        className="form-checkbox border-0 rounded text-blueGray-700 ml-1 w-5 h-5 ease-linear transition-all duration-150"
-                      />
-                      <span className="ml-2 text-sm font-semibold text-blueGray-600">
-                        Remember me
-                      </span>
-                    </label>
-                  </div>
+                
+               
 
                   <div className="text-center mt-6">
                     <button
@@ -69,7 +45,7 @@ export default function Login() {
                       type="button"
                       onClick={handleUserLogin}
                     >
-                      Sign In
+                      Reset Password
                     </button>
                   </div>
                 </form>
@@ -79,10 +55,7 @@ export default function Login() {
               <div className="w-1/2">
                 <a
                   href="#pablo"
-                  onClick={(e) => {
-                    e.preventDefault();
-                    router.push("/auth/forgetpassword");
-                  }}
+                  onClick={(e) => e.preventDefault()}
                   className="text-blueGray-200"
                 >
                   <small>Forgot password?</small>
@@ -103,4 +76,4 @@ export default function Login() {
   );
 }
 
-Login.layout = Auth;
+ForgetPassword.layout = Auth;