+++ pages/auth/index.js
@@ -1,39 +1,57 @@
 import React from "react";
-import Link, { Router } from "next/link";
-import { useRouter } from "next/router";
+import Link from "next/link";
 
 // layout for page
 
 import Auth from "layouts/Auth.js";
 
 export default function Login() {
-  const router = useRouter();
-
-  const handleUserLogin = async (e) => {
-    router.push("/admin/dashboard");
-  };
-
   return (
     <>
       <div className="container mx-auto px-4 h-full">
         <div className="flex content-center items-center justify-center h-full">
           <div className="w-full lg:w-4/12 px-4">
             <div className="relative flex flex-col min-w-0 break-words w-full mb-6 shadow-lg rounded-lg bg-blueGray-200 border-0">
+              <div className="rounded-t mb-0 px-6 py-6">
+                <div className="text-center mb-3">
+                  <h6 className="text-blueGray-500 text-sm font-bold">
+                    Sign in with
+                  </h6>
+                </div>
+                <div className="btn-wrapper text-center">
+                  <button
+                    className="bg-white active:bg-blueGray-50 text-blueGray-700 font-normal px-4 py-2 rounded outline-none focus:outline-none mr-2 mb-1 uppercase shadow hover:shadow-md inline-flex items-center font-bold text-xs ease-linear transition-all duration-150"
+                    type="button"
+                  >
+                    <img alt="..." className="w-5 mr-1" src="/img/github.svg" />
+                    Github
+                  </button>
+                  <button
+                    className="bg-white active:bg-blueGray-50 text-blueGray-700 font-normal px-4 py-2 rounded outline-none focus:outline-none mr-1 mb-1 uppercase shadow hover:shadow-md inline-flex items-center font-bold text-xs ease-linear transition-all duration-150"
+                    type="button"
+                  >
+                    <img alt="..." className="w-5 mr-1" src="/img/google.svg" />
+                    Google
+                  </button>
+                </div>
+                <hr className="mt-6 border-b-1 border-blueGray-300" />
+              </div>
               <div className="flex-auto px-4 lg:px-10 py-10 pt-0">
-                <div className="text-blueGray-400 text-center mb-3 font-bold"></div>
+                <div className="text-blueGray-400 text-center mb-3 font-bold">
+                  <small>Or sign in with credentials</small>
+                </div>
                 <form>
                   <div className="relative w-full mb-3">
                     <label
                       className="block uppercase text-blueGray-600 text-xs font-bold mb-2"
                       htmlFor="grid-password"
                     >
-                      Phone number
+                      Email
                     </label>
-
                     <input
-                      type="number"
+                      type="email"
                       className="border-0 px-3 py-3 placeholder-blueGray-300 text-blueGray-600 bg-white rounded text-sm shadow focus:outline-none focus:ring w-full ease-linear transition-all duration-150"
-                      placeholder="Phone number"
+                      placeholder="Email"
                     />
                   </div>
 
@@ -67,7 +85,6 @@ export default function Login() {
                     <button
                       className="bg-brand-purple text-white active:bg-blueGray-600 text-sm font-bold uppercase px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 w-full ease-linear transition-all duration-150"
                       type="button"
-                      onClick={handleUserLogin}
                     >
                       Sign In
                     </button>
@@ -79,10 +96,7 @@ export default function Login() {
               <div className="w-1/2">
                 <a
                   href="#pablo"
-                  onClick={(e) => {
-                    e.preventDefault();
-                    router.push("/auth/forgetpassword");
-                  }}
+                  onClick={(e) => e.preventDefault()}
                   className="text-blueGray-200"
                 >
                   <small>Forgot password?</small>