+++ CHANGELOG.md
@@ -13,8 +13,8 @@
     - Then, you search in your whole project for `blueGray-500` and replace it with `blueGray-400`
     - Then, you search in your whole project for `blueGray-600` and replace it with `blueGray-500`
     - Then, you search in your whole project for `blueGray-700` and replace it with `blueGray-600`
-    - Then, you search in your whole project for `blueGray-800` and replace it with `blueGray-700`
-    - Then, you search in your whole project for `blueGray-900` and replace it with `blueGray-800`
+    - Then, you search in your whole project for `brand-purple` and replace it with `blueGray-700`
+    - Then, you search in your whole project for `blueGray-900` and replace it with `brand-purple`
   - For the colors, the only change that we made, is the fact that we've added all Tailwind CSS colors to our `tailwind.config.js` files, and inside our product, all `{type}-blue-{number}` classes were renamed to `{type}-lightBlue-{number}`
   - For the colors, the only change that we made, is the fact that we've added all Tailwind CSS colors to our `tailwind.config.js` files, and inside our product, all `{type}-green-{number}` classes were renamed to `{type}-emerald-{number}`
 - `lg:bg-transparent` is not working anymore, so we've changed it with `lg:bg-opacity-0`